split macro file
